debezium/dbz#1812 Move general purpose util classes to debezium-util package
