+++ debezium-connector-mongodb/pom.xml
@@ -53,6 +53,12 @@
         </dependency>
         <dependency>
             <groupId>io.debezium</groupId>
+            <artifactId>debezium-util</artifactId>
+            <type>test-jar</type>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>io.debezium</groupId>
             <artifactId>debezium-connect-plugins</artifactId>
             <type>test-jar</type>
             <scope>test</scope>
