+++ debezium-storage/debezium-storage-tests/pom.xml
@@ -48,6 +48,11 @@
         </dependency>
         <dependency>
             <groupId>io.debezium</groupId>
+            <artifactId>debezium-util</artifactId>
+            <type>test-jar</type>
+        </dependency>
+        <dependency>
+            <groupId>io.debezium</groupId>
             <artifactId>debezium-ddl-parser</artifactId>
             <scope>test</scope>
         </dependency>
