+++ debezium-util/src/main/java/io/debezium/util/ElapsedTimeStrategy.java
