+++ debezium-util/src/main/java/io/debezium/util/HashCode.java
