+++ debezium-util/src/main/java/io/debezium/util/HexConverter.java
