+++ debezium-util/src/main/java/io/debezium/util/Iterators.java
