+++ debezium-util/src/main/java/io/debezium/util/Joiner.java
