+++ debezium-util/src/main/java/io/debezium/util/MathOps.java
