+++ debezium-util/src/main/java/io/debezium/util/NumberConversions.java
