+++ debezium-util/src/main/java/io/debezium/util/Sequences.java
