+++ debezium-util/src/main/java/io/debezium/util/VariableLatch.java
