+++ debezium-util/src/test/java/io/debezium/text/PositionTest.java
