+++ debezium-util/src/test/java/io/debezium/text/TokenStreamTest.java
