+++ debezium-util/src/test/java/io/debezium/text/XmlCharactersTest.java
