+++ debezium-util/src/test/java/io/debezium/util/HexConverterTest.java
